Moved resource strings AppWideResources->VestaImporter
